Initial Commit
I’ve known Ed Spicer almost my entire life. He’s a legendary teacher,
and an amazing advocate for children’s literature. Someone who’s a kid
at heart in the best way.

In his retirement, he’s revamping his personal website. Publishers,
friends of his from conferences, and anyone with an interest in
children’s literature is the audience.

As the name of the site is “Spicy Reads”, I decided to use a color
palate that is lively and loosely based off of peppers.

As most of Ed’s endeavors will happen off-site (his Medium blog,
Twitter account, YouTube page), this is mostly a hub to offer an
introduction to those outlets.

As this is a site dedicated to children’s literature, I’m going to try
to keep things as playful as possible.