Contact and Copy
Ed gave me some copy for his video section and also for his contact
section.
BIG NEWS: I used a table for the first time on a site. I know they’re
kinda the inside joke of modern web dev but, ya know, it really worked
when it came to lining up the icons and the explanations on the contact
section.