Home Page and CTA
My good friend Adam Salois is a designer and he gave things a look over
and gave me some feedback.

I’m turning all the links into buttons to make the calls to action
clearer.

Also the home page was a little clunky with the title placard. To fix
that I put the title in the nav.

The title and the navbar stack on mobile.